✏️ Moved steps files out of UI folder
